requirements.txt
# HunyuanWorld ComfyUI Integration Dependencies

# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=8.0.0

# Computer Vision & Image Processing
opencv-python>=4.5.0
scikit-image>=0.19.0

# 3D Processing & Mesh Handling
trimesh>=3.9.0
pymeshlab>=2022.2
open3d>=0.15.0

# Mathematical & Scientific Computing
scipy>=1.7.0

# Networking & Model Loading
requests>=2.25.0
tqdm>=4.60.0
huggingface-hub>=0.10.0

# Optional: Enhanced Model Support
# diffusers>=0.20.0  # For diffusion models
# transformers>=4.20.0  # For transformer models
# accelerate>=0.20.0  # For model acceleration

# Optional: Better 3D Format Support
# pygltflib>=1.15.0  # For GLB/GLTF export
# pyfbx>=0.1.0  # For FBX export (if available)

# Optional: Advanced Image Processing
# realesrgan>=0.3.0  # For image enhancement
# basicsr>=1.4.0  # Image restoration

# Development Dependencies (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0